Update filament for GPU from Imagination Technologies

+++ filament/src/materials/ssao/depthUtils.fs
@@ -28,11 +28,11 @@ highp float linearizeDepth(highp float depth) {
     return (depth * p[2].z + p[3].z) / max(depth * p[2].w + p[3].w, preventDiv0);
 }
 
-highp float sampleDepth(const highp sampler2D depthTexture, const highp vec2 uv, float lod) {
+highp float sampleDepth(const mediump sampler2D depthTexture, const highp vec2 uv, float lod) {
     return textureLod(depthTexture, uvToRenderTargetUV(uv), lod).r;
 }
 
-highp float sampleDepthLinear(const highp sampler2D depthTexture,
+highp float sampleDepthLinear(const mediump sampler2D depthTexture,
         const highp vec2 uv, float lod) {
     return linearizeDepth(sampleDepth(depthTexture, uv, lod));
 }
